Prepare io/console for move to gem.
This commit makes the following changes:

* Move JRuby-specific console logic to io/console/jruby
* Add console.rb to load appropriate impl for the current engine
* Adjust tags to match passing tests
* Add test_io_console to MRI stdlib suite

See ruby/io-console#11
