+++ lib/ruby/stdlib/io/console/jruby/bsd_console.rb
