+++ lib/ruby/stdlib/io/console/jruby/common.rb
