+++ lib/ruby/stdlib/io/console/jruby/linux_console.rb
