+++ lib/ruby/stdlib/io/console/jruby/stty_console.rb
