+++ lib/ruby/stdlib/io/console/jruby/stub_console.rb
