+++ test/mri/excludes/TestIO_Console.rb
@@ -1,17 +1,18 @@
 # frozen_string_literal: false
-exclude :test_cooked, 'depends on unimplemented PTY functionality'
-exclude :test_echo, 'depends on unimplemented PTY functionality'
+# exclude :test_cooked, 'depends on unimplemented PTY functionality'
+# exclude :test_echo, 'depends on unimplemented PTY functionality'
 exclude :test_getpass, 'not portable'
-exclude :test_iflush, 'depends on unimplemented PTY functionality'
-exclude :test_ioflush, 'depends on unimplemented PTY functionality'
-exclude :test_ioflush2, 'depends on unimplemented PTY functionality'
-exclude :test_noecho, 'depends on unimplemented PTY functionality'
-exclude :test_noecho2, 'depends on unimplemented PTY functionality'
-exclude :test_oflush, 'depends on unimplemented PTY functionality'
-exclude :test_raw, 'depends on unimplemented PTY functionality'
-exclude :test_raw!, 'depends on unimplemented PTY functionality'
+# exclude :test_iflush, 'depends on unimplemented PTY functionality'
+# exclude :test_ioflush, 'depends on unimplemented PTY functionality'
+# exclude :test_ioflush2, 'depends on unimplemented PTY functionality'
+# exclude :test_noecho, 'depends on unimplemented PTY functionality'
+# exclude :test_noecho2, 'depends on unimplemented PTY functionality'
+# exclude :test_oflush, 'depends on unimplemented PTY functionality'
+# exclude :test_raw, 'depends on unimplemented PTY functionality'
+# exclude :test_raw!, 'depends on unimplemented PTY functionality'
 exclude :test_raw_minchar, 'depends on unimplemented PTY functionality'
 exclude :test_raw_timeout, 'depends on unimplemented PTY functionality'
-exclude :test_setecho, 'depends on unimplemented PTY functionality'
-exclude :test_setecho2, 'depends on unimplemented PTY functionality'
-exclude :test_winsize, 'depends on unimplemented PTY functionality'
+exclude :test_set_winsize_invalid_dev, 'depends on unimplemented PTY functionality'
+# exclude :test_setecho, 'depends on unimplemented PTY functionality'
+# exclude :test_setecho2, 'depends on unimplemented PTY functionality'
+# exclude :test_winsize, 'depends on unimplemented PTY functionality'
