+++ test/mri/io/console/test_io_console.rb
@@ -355,7 +355,7 @@ defined?(IO.console) and TestIO_Console.class_eval do
     noctty = [EnvUtil.rubybin, "-e", "Process.daemon(true)"]
   when !(rubyw = RbConfig::CONFIG["RUBYW_INSTALL_NAME"]).empty?
     dir, base = File.split(EnvUtil.rubybin)
-    noctty = [File.join(dir, base.sub(/ruby/, rubyw))]
+    noctty = [File.join(dir, base.sub(RUBY_ENGINE, rubyw))]
   end
 
   if noctty
